src/requirements.txt
google-api-core>=2.24.2
google-api-python-client>=2.169.0
google-auth>=2.39.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.2
python-dotenv>=1.0.0
openai>=1.77.0
sounddevice>=0.4.6
numpy>=1.24.0
python-dotenv>=1.0.0
pydantic>=2.11.4
click>=8.1.7
rich>=13.7.0 